# Conflicts:
#	build.gradle

+++ api-core-cms/src/main/java/gov/ca/cwds/data/persistence/cms/BaseSubstituteCareProvider.java
@@ -22,9 +22,6 @@ import gov.ca.cwds.data.std.ApiPersonAware;
 public abstract class BaseSubstituteCareProvider extends CmsPersistentObject
     implements ApiPersonAware {
 
-  /**
-   * Base serialization version. Increment per change.
-   */
   private static final long serialVersionUID = 1L;
 
   @Id
@@ -32,7 +29,7 @@ public abstract class BaseSubstituteCareProvider extends CmsPersistentObject
   protected String id;
 
   @Column(name = "ADD_TEL_NO")
-  protected BigDecimal additionalPhoneNumber;
+  protected Long additionalPhoneNumber;
 
   @Type(type = "integer")
   @Column(name = "ADD_EXT_NO")
@@ -191,7 +188,7 @@ public abstract class BaseSubstituteCareProvider extends CmsPersistentObject
   /**
    * @return the additionalPhoneNumber
    */
-  public BigDecimal getAdditionalPhoneNumber() {
+  public Long getAdditionalPhoneNumber() {
     return additionalPhoneNumber;
   }
